Added column deletion on Db:create
Db:create will now delete columns, if they are present in the current database,
but not in the new schema. All data in said columns will be lost. The change was
necessary as former versions of the library added options as columns, which was
rectified but made existing databases unreadable to the library.