Mudlet now luads LuaGloba.lua, but nothing inside it
Mudlet can now locate and load LuaGlobal.lua, but nothing inside it
works.

Additionally, Mudlet now has to be built as a .app bundle on a Mac
because LuaGlobal will only be found that way - which is fine as .app's
are the preferred Mac distribution method, and it's no extra resources
to make one at compile time.