Fixup: Add new distinct icon for Room Exits' Dialog
Was unable to insert the icon in previous commit as it was tucked away in
a different prototyping branch and not then accessible!

Modify some tool-tip code to handle unnamed exit rooms better, and correct
erroneous application of tool-tips to the frame for the group of controls
for each exit.

Also adjust spacing on some bits for consistency with rest of file.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>