BugFix/Enhancement: Custom Line context menu/GUI
The ability to draw new custom exit lines had been lost in the previous
SF Git master.  This reverts the GUI changes previously made and enhances
the results.  The former dialog is now more intuitive.  It is now possible
to (repetitively) undo line segments placed during the drawing operation
and to modify the characteristics of the line both during drawing and
subsequently.  It is also now possible to modify an existing line by adding
or removing segments also via an enhanced context menu.

Also to aide editing and drawing new lines, a bright yellow target is drawn
around the destination room!  When editing an existing line, the currently
selected custom line vertex is also drawn in yellow rather than orange so
the user knows which point will be modified if a vertex is added or
removed.  The help message used during some 2D mapper operations including
some custom line actions is now positioned centrally at the bottom of the
screen (in a somewhat smaller font size) so it no longer clashes with the
map info display.  It now also appears at the correct point in drawing
custom lines whereas previously it only appeared after the first point was
created by left clicking on the map.

A minor edit to TLUaInterpreter::addCustomLine() now adds support to all
five standard line styles available in current Qt version.

An Icon for the new dialog (custom_lines_properties.ui) has been added to
the resources file and a related icon also created for the custom_lines.ui.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>