Enhancement: Centralised control of Mudlet version.
Setting the (pre-existing) VERSION qmake variable to a 1.2.3
(major.minor.revision) form and the BUILD variable to nothing (for a
release build) or some non-empty string (possibly with a leading '-'
to act as a separator) creates version information that is uniformly
presented in several places.  This includes on the splash screen, in the
"About Mudlet" dialog, in some communications with the MUD server, and
additionally now within the Lua system for script writers.

In addition I have updated and tidied some detail of, and added myself to
the listed Mudlet contributors in, the "About Mudlet" dialog.

In reviewing the other tab of the "About" dialog that gives a reference to
the Licensing terms it became clear that its claim that Mudlet was
Licensed to GPL version 3 or greater was inconsistent with the text header
of most of the source files which explicitly state version 2 or greater.
To rationalise this I changed the dialog to also say version 2 or later and
then provide the full and exact text of the "current" version 3.  I also
changed that tab's type so that the hyperlinks are now functional.

Whilst paying attention to the GPL text it became clear that some tweaks
to the application start up process was needed to bring Mudlet's behaviour
to that which users of GPL software expect.  This includes responding to
the standard "h / help" and "v / version" arguments when Mudlet is started
from a command line and displaying some specific pieces of text during
normal start-up.  To placate a user that does not wish to be bothered with
the slightly longer splash-screen display process Mudlet now also accepts a
"q / quiet" argument to hide it altogether, or, as is standard for
QtSplashScreen they can also click on it to achieve a similar result.  To
cover the hiatus during start-up the cursor will now change to the OS
standard "Wait" one until the main event loop starts.

THE CORRECTION IN LICENCE TEXT NEEDS TO BE CHECKED BY HEIKO AS IT IS DONE
IN HIS NAME. 8-)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>