Merge Prepare: Rearrange project file to make it more merge-able
In preparation for use as Pull Request #32 on:
http://github.com/Chris7/Mudlet2/mudletDev reorganise src.pro so that it
more closely matches my branch
http://github.com/Slysven/Mudlet2/mudletDev_fileLocation
which is lined up as Pull Request #31 on the same one of Chris's branch.

There should not be any change in overall functionality (the qmake'ed
Makefile should produce the same executable both before and after this
commit).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>