Tweaks: some small changes to custom line drawing code
Change the circles around custom line points when selected to be solid
instead of dotted - looks better under a wider range of room and exit
size settings.  Simplify the target drawing code to indicated the target
room of a custom exit line - also looks better.

Added code for the first segment of a custom exit line, was previously
explicitly not handled because of the extra code that it would require.

Status bar messages for various custom line exit situations extended to
cover corner cases.

NB. StatusBar is NOT currently showing on at least Qt 5.1 / GNU/Linux
platform.  A code lines or two is needed in mudlet.cpp to generate it.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>