BugFix: Fix-up various iconic issues.
In four places (one a twin entry) the resource path was given as:
":/icons/icons/<icon>.png" when it should have been: ":/icons/<icon>.png",
additionally as one of these uses was for a dialog box where it closed
that box and committed any changes made with no option to cancel without
undoing them I feel it would be more appropriate to use an "apply" icon
(a green tick) rather than the originally intended "dialog-close" (a
white cross on a red filled in circle).

The main application was assigned the 16x16 icon but this appears blurry
on the system task switching control (on at least Linux and Windows)
platforms.  Using the 48x48 icon appears much better and has been
changed in mudlet::mudlet().  The central debug console was not assigned
an icon so in the system task switcher it appeared as a plain icon
(at least in Linux) - I have created a suitable icon (putting a quarter
size bug on the bottom right quarter of the mudlet 48x48 icon).
Similarly I have created one for the profile_preferences.ui form with
a superimposed spanner and actually used the mudlet_notepad one that was
already present but not used for the profile notes editor.  This makes
if a bit easy to tell which window is which when swapping between them
by keyboard!

When the profile_preferences.ui has the "Use Mudlet on a netbook"
feature set, the icon that appears on the Main Toolbar to toggle in
and out of "fullscreen" mode is the Oxygen "dialog-cancel" one, however
there is a specific icon "view-fullscreen" which more closely relates
to the function of this control, I have included it and changed to that
one.  As the control is actually a toggle I also have included the
corresponding "view-restore" image as well which it might be possible
to code to use...

On the "replay" toolbar the "replay speed down" icon was spelt as
"port.png" rather than "import.png" so the downward pointing double
arrow was not shown, leaving an empty button in some situations.

Within the Trigger Editor it has been commented by users that the icon
labelled "save" and featuring a disc symbol does not actually write
anything in a permanent manner and though the tooltip does warn of this
I believe a more coherant use could be made by replacing some of the
icons.  Building on the top tool bar "item" and "group" use of a page
and a folder respectively I have used the executing and error/stop
sub-icons from the oxygen "printer-printing" and "printer-error" images
and placed them on top of a page image to replace the
tag-checkbox-checked amd tag-checkbox for the item enabled/disabled
treeWidget icons, a second pair rendered in grey scale becomes the
ancestor disabled pair and one with a "bug" to the same scale represents
code that cannot be Lua compiled.  Combining both executing and error
sub-icons ontop of a pair of circular arrows on top of a page replaces
the activate/deactive toolbar button and the circular executing sub-icon
recoloured from green to orange with an arrow indicating the affixing of
the sub-icon onto a page becomes the slightly previously misleading
"save" (code) button next to it on the top toolbar.  In the special case
of a "circular" Alias a variant of this in red with a red cross over the
arrow is used as it indicates that the "code" cannot be affixed to the
"page", which does represent the fact that the page can't be saved until
it is fixed.  As a nice side effect, that extends the page metaphor, is
the oxygen "document-import" and "document-export" icons that,
particularly in the export case, emphasises that only the selected page
(or folder) IS what is exported.  Also, as we no longer use the oxygen
"save" (disk) icon to write the code into the system, we can now use
it for the save profile (replacing the double disc "save all" icon)
which sits more nicely alongside the "save as" (disc with a pencil
writing a new label) icon that is already used.  Finally, the delete
item used the oxygen "edit-delete-shred" image but that uses a
"crumpled" page which although perhaps appropriate does look visually
at odds with the clean straight edges of the other "page" icons.
As there didn't seem to be a "document-delete" image in the Oxygen set
I had to make one to match.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>