Use QStringLiteral more liberally when possible.
In most places where an Icon is specified a fixed QString that refers too
the location in the resource file is needed.  QStringLiteral() is a macro
that allows the implict conversion from const char to a QString to be done
at compile time rather than run-time.  Also the result can be stored in
read only memory.  Also it helps to explicitly mark a string that is NOT
to be translated by tr() which will help in the future if we ever attempt
to translate Mudlet into languages other than en-US (yippee - as an
Englishman I won't have to put up with color rather than colour
everywhere!)  If anyone is brave enough to set the
Q_NO_CAST_FROM_ASCII/Q_NO_CAST_TO_ASCII macros they will get to see how
many strings will have to be either wrapped in QStringLiteral or tr() -
not counting error messages in the Lua stuff in std::string form that will
have to be redone into QString until the point of conversion to/from utf-8
into or out of the Lua subsystem so that it can also be translated!!!