Fixes: Some dialog icon issues
The Window Icons were not being used for profile_preferences dialog or
for the Central Debug console.  In the former it transpires the icon
was loaded from a relative file reference and not from the resource
file (and also for the icon for mudlet_custom_exit_properties) file;
and in the latter it was being set by code from the mudlet class which
just did not work - as it is being used as the central widget inside
a QMainWindow widget and THAT is what has to have the icon set!

Also, fix the fact that the profile preference was not opened in the
first tab but the third.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>