Re-factor: some error messages during Lua startup + prep. 2 fnctns for i18n
As first of a few commits to reattempt to include my Unix installation
system that supports static Lua code (external Mudlet Lua and Geyser) files
I spotted that the error message generation code did not handle the case
where Lua error messages were NOT available in the same way as when they
were.  This commit improves the code and WILL work in the future for
multilingual code where we use the tr() functionality for user facing
translatable code and QStringLiteral() wrapped const strings for internal
strings NOT seen by the user.  As such TLuaInterpreter::initLuaGlobals()
and TLuaInterpreter::loadGlobal() will support interaction with a
Lua-subsytem using Utf-8 encoded strings which are a super-set of ASCII.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>