Enhancement: permit Lua files to be away from subdirectory under executable
For *nix systems in particular (but can be applied to other environments)
the core Lua files used by Mudlet (and those for the Geyser system) will
ideally be placed in a read-only shared location away from the binary
(executable) Mudlet application.  This commit includes sufficient
amendments to both cmake and qmake project files to perform the traditional
make install (when invoked via sudo) to install the binary to typically
/usr/local/bin and the lua files to /usr/local/share/mudlet/lua and for
the executable to find the latter when it is run.  The location of the Lua
loader of all the other files is validated at the start of use of a profile
and if not valid the user is permitted to seek out the required
LuaGlobal.lua file.  If the file is NOT found the user is warned that
Mudlet will be very restricted and they are advised to seek assistance, if
the file IS found than it's location is stored in the per user
configuration data for Mudlet so the user will not be asked again, unless
the file is removed.

To permit developers to modify and test the set of Lua files included with
the rest of the source code, a check box has been added using the previous
commit's "runtime debug framework" to force any subsequently loaded
profiles to use the Lua files located in the source location rather than
the recorded "system" place.  This setting is also retained between
invocations of Mudlet.  To prevent users from breaking their installation
if they are running a packaged binary from a distribution this override
control is disabled in the absence of a header file included in the source
Lua files which is written by the project file to provide the information
about the absolute path of said files and which is difficult to establish
consistently given the range of compiler systems that may be used and how
they handle the __FILE__ pre-processor macro (though that gives the name of
the file being compiled or included at compile time it may not be an
absolute one!)

The default location set in the top level project file is compiled into the
application so it behoves packager to set it to the correct value for their
intended target audience and why this commit also tries to configure for
a correct INSTALLATION system.  Both CMake and qmake project files are
functional for Linux, some tweaks may be required for Mac Bundles and no
attempt has been made to provide an installer for Windows.  As such this
would benefit from the attentions of those who develop on those platforms.


Signed-off-by: Stephen Lyons <slysven@virginmedia.com>