Discuss clang-format options.
This isn't meant to be merged, but rather, be a discussion about which
options people want to use. I've added comments next to each formatting
option that clang-format supports and that we specify. Many of them are
fairly obvious choices and some of the settings I'm ignoring for now,
either for future discussion or because they are for fine tuning later and
not important at the moment.

For things I want to call out as potentially contentious, I'll have the
comments start with **'s.