Fixup: unsquash cust.line. form "cancel" button + swap pressed() to clicked()
Removed the fixed constraint on the cancel button on the custom_lines.ui
form - as on some systems it was too narrow - it now fills the available
width of the dialog.  I clarified the tool-tip for this button slightly
to make the information more closely match reality

Following comments from Vadim, changed code associated with all buttons on
this form to use the clicked() signal rather than the previous pressed()
one.

Also correct a spelling mistake in several error messages.

Noting the same usage on the room_exits.ui form, performed the SAME change
from pressed() to clicked() signals there.

Also to avoid any squishing of the four buttons on the bottom of THAT
dialog set uniform minimum sizes for them.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>