Debug code fix: for two bugs in debugging code
First bug was one that printed out a copy of the current chunk of the
replay for EACH character in that chunk - fix was to move the print line
to be outside of the for() loop that examines each character.

Second bug was an off-by-one one that over-estimated the length of the
chunk of data from the replay file.

Although this commit temporarily enables the relevant debug lines I intend
to comment them back out again at the end of the series of commits!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>