Enhance: Update Replay time display every second rather than every chunk
For replays where input occurs less often than once a second the time
display was only updated when each chunk was read with the result that the
time display could be static for long periods.  This commit increments the
display by one second every second divided by the speed factor.  Due to
the way it is implemented it is somewhat approximate but the displayed time
is reset to the right value when each chunk is processed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>