Enhance: Add lua commands to get replay status, control speed & abort replays
As it is possible to initiate a replay from with a user Lua script it does
not seem unreasonable to also provide command to abort a replay, to obtain
various bits of status information about such a replay and to provide a
means to control the playback speed in the same way as the toolbar does.
This commit provides all of these features.

Debug displays of chunk data have been commented out with an explanatory
comment!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>