Enhance: Cache CSS style data for TTextEdit::copySelectionToClipboardHTML()
Implements the previously mentioned CSS cache - analyse clipboard
selection's CSS style and replace each instance in the HTML <body> with
a class reference to the same in the <head> part of the HTML document.  A
test case of a log from over an hours worth of play on a MUD reduced a file
of size 1.7 to just under 0.9 MBytes with no appreciable delay.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>