Revision: force LFs with <br> in TTextEdit::copySelectionToClipboardHTML()
Following a request both literal line-feeds AND <br> tags are used to mark
an actual end of line in the HTML produced by the Console/MiniConsole
copy to HTML function.  To prevent double-LFs from occurring the CSS
"white-style" mode has been changed from "pre-wrap" to "nowrap".  This
means that the source HTML is not all on one line when viewed in a normal
text editor as the parse of a web-browser should now ignore LFs in the
HTML.  However it also means that tabs may not work...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>