Enhance: Update Replay time display every 1/10 second not every chunk
For replays where input occurs less often than once a second the time
display was only updated when each chunk was read with the result that the
time display could be static for long periods.  This commit increments the
display by one second every second divided by the speed factor.  A previous
attempt to do this caused some "jumping/sticking" in the display because
the displayed time was advanced only once a scaled second, which meant the
displayed time did not seem to increase uniformly as it was reset to the
right value when each chunk is processed.  By decreasing the amount the
displayed time was incremented to a scaled tenth of a second the
perturbations is reduced considerably.

As the used time format {"hh:mm:ss"} is the default for the
QTime::toString() method used to generate the text to display on the Replay
Time display widget we do not need to have a const QString constant to hold
the time display format to use as the record of the formatting we want.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>