Enhance: Add sysReplayEvents to notify user scripts of asynchronous changes
The replay system can only handle one replay at a time, and although a user
script can initiate a replay it previously could not know about the ending
of that replay.  Also if a replay is taking place in one profile, another
one will get informed that the system is busy with a replay but cannot know
without repeatedly trying to start it's own when the system is available.
This commit defines several sysReplayEvents, with additional arguments as
follows:

"myReplayStart" - a replay that this profile has requested, either from a
    script or from the main tool bar has been opened SUCCESSFULLY and is
    now running.  Note that a failure to start the reply will be reported
    back in an error return to a lua script that tries to start one so does
    not need an event.

"otherReplayStart" - a replay has been started in another profile and so
    replays cannot currently be started by this profile.

"myReplayOver" - the replay that this profile had started has finished.

"otherReplayOver" - the replay that another profile had started has
    finished.  The replay system is thus now available to this profile.

"replaySpeedChange", (float)new_replay_rate, (float)old_replay_rate - not
    much use currently but notifies the profile running a replay that the
    replay speed has just been changed, the following arguments are the new
    and the previous rates (x0.125 = 1/8 to x128).  A lua command
    getReplayStatus() will follow in another commit to return a current
    and total replay times so combining this with that data may give useful
    information if a script wants to estimate replay timings.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>