Enhance: Add magic to replay files, makes replay system more robust
Until now the Mudlet replay files had the generic ".dat" file extension and
there was no protection against the user loading a file that is NOT a
replay file.  Given that the file includes embedded values to control the
amount to read from the file using a time sent to a QTimer in another
embedded value this seems unwise.  This commit adds some data to the start
of a replay file that serves to unique identify it (and adds a version code
so that if the format is changed in the future old formats can still be
identified correctly!)  It also changes the extension to ".mreplay" to
stand for Mudlet replay, moving away from a three letter string as an
extension as that restriction went away in operating systems probably way
older than Mudlet ever supported.  Old format replays will still be
acceptable, but the user will have to choose to use those by changing the
file type filter from the file requester dialog that is called when a
replay is requested from the main toolbar button.

The use of "dd-MM-yyyy#hh-mm-ss" as the time format to name the replay file
may look pretty but is unhelpful in sorting as an alphabetical sort of the
directory contents will group by day of month, month and then year etc.
This is OK provided the date-time meta-data is preserved but if such files
are archived or copied that date-time information can be lost and then
picking out files by age can become very difficult if the listing takes
more than one screen-full.  Putting the most significant date-time item
first, as is common in some world regions, makes an alphanumeric sort
produce the same results as a date-time meta-data sort and preserves the
sort order if the latter data is corrupted.

In addition, by structuring some elements in the header in a particular
layout it makes it possible to determine some characteristics about the
file from the command line - this is particularly relevant on *nix
platforms which offer the file(1) command using data parsing routines
configured by magic(5).

The differences now mean instead of receiving the lowest common denominator
of file type information, the generic "data" type:
'$> file ./WoTMUD_Jomin/log/23-02-2014#12-31014.dat
./WoTMUD_Jomin/log/23-02-2014#12-31014.dat: data'

one instead can get:
'$> file ./WoTMUD_Jomin/log/2014-09-12#17-28-11.mreplay
./WoTMUD_Jomin/log/2014-09-12#17-28-11.mreplay: Mudlet replay file, version 2, started: Fri Sep 12 17:28.11 2014(UTC), 15.106(s) long, profile: "WoTMUD_Jomin"'

As it is possible to initiate a replay from with a user Lua script it does
not seem unreasonable to also provide commands to:
* abort a replay
* obtain various bits of status information about such a replay
* control the playback speed in the same way as the toolbar does.

Some existing class members/methods have been modified for clarity or other
reasons:
* TConsole::mReplayFile     ==> TConsole::mRecordFile
* TConsole::mReplayStream   ==> TConsole::mRecordStream
  above two are associated with making a recording not with its playback
* TConsole::mRecordReplay   ==> TConsole::mIsRecording
  to match corresponding cTelnet::mIsReplaying.
* cTelnet::readPipe()       ==> cTelnet::slot_readPipe()
  to emphasise it is a slot method.
* QTime cTelnet::timeOffset ==> QDateTime cTelnet::mRecordLastDateTimeOffset
  instead of measuring the elapsed time directly we now save the point in
  time when the last chunk of data item was recorded, this allows for the
  unlikely but possible case of records of longer than a day or for when
  DST changes - both of which a QTime can not handle.
* mudlet::mpReplayTimer     ==> mpReplayDisplayTimer
  to differentiate it from the new mpReplayChunkTimer, the first is
  associated with updating the time display periodically every
  (second/speed) and the second with loading the next chunk after the
  right time period.
  Both are included in the class header so they can be accessed /
  controlled from outside of the class which is not possible if they are
  declared as class globals as the first had been in the past.
* cTelnet::abortReplay()    ==> cTelnet::abortReplaying()
  to match corresponding TConsole::abortRecording().
* cTelnet::mLoadingReplay   ==> cTelnet::mIsReplaying
  to match corresponding TConsole::mIsRecording.

Also:
mLogFileName, local to TConsole::slot_toggleReplayRecording() MASKS a
variable with the same name declared in the class header, it has been
renamed to recordingPathFile.  Another local variable of the same name in
TConsole::slot_toggleLogging() had the same effect - the local definition
there has now been removed so that that method will use the original member
of the class instance after all.

In TConsole.h the member mpCommandLine has been taken out of the grouping
of public methods and re-positioned further down with the other public
members.  Similarly the cTelnet::postMessage( QString ) method has been
moved out from the public members and up to the bottom of the other public
methods.

When making a recording Mudlet now should detect a failure to write out a
chunk of replay data completely and will abort, try to clean up the end of
the recording if possible, and report an error in the main profile console.

During the loading/playing of replays the data was being loaded into a
static fixed size char array { cTelnet::mLoadBuffer } of 100 Kilo-Bytes!
However as we know the size of each chunk before it is read from the file
it can easily be changed to a QByteArray that can be enlarged as required
if it is not big enough to hold a particular chunk of replay data.  In
a quick test the array only grow to a few thousand bytes and will only
slightly slow things down for each chunk read for which the buffer is
undersized until the maximum size needed is reached.

Although not marked with any licence terms I hereby state that the "magic"
file entry added to the INSTALL file is, for the avoidance of any confusion
to be considered as licensed as "GPLv2+" - if that is not suitable then
please contact me for alternative options:

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>