BugFix: Fix underlines AND italics in HTML copy.
The CSS declaration for underline (and overline and strike-through) is
introduce by the words "text-decoration" NOT "font-decoration", this commit
fixes that error. https://bugs.launchpad.net/mudlet/+bug/1130450 raised
this issue and is solved by it.

Also the CSS declaration for a font-style of sloping text derived from a
font that is constructed for that sloping text is "italic" singular NOT
"italics" plural.  N.B. The other alternative to "normal" is "oblique"
which specifies sloping text derived from a font that is upright but
converted to sloping by algorithm rather than definition...!  Correcting
this solves: https://bugs.launchpad.net/mudlet/+bugs/1102360

This is a bare-bones commit that addresses these issues only. Further work
is planned to produce Html that is both actually valid and more compact,
there are still other issues to be addressed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>