Enhance: Enable replay to take place at less than unity speeds
This commit uses negative values of the mudlet class member mReplaySpeed to
indicate speeds that are the reciprocal of the integer number.  It tweaks
the display of the speed to show a fractional value rather than a decimal
one in the limited cases (1/2, 1/4, 1/8) that are coded for.  As the lowest
supported speed is altered by this commit it also seemed reasonable to set
a reasonable upper limit of 128 so that one second of replay corresponds to
just over two minutes of real-time in the original recording.

At the same time the infrastructure to support translation of the widgets
that contain text is implemented for future internationalisation.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>